# swagger.yaml
openapi: 3.0.0
info:
  title: Runway Kicks API
  version: 1.0.0
  description: API documentation for the Runway Kicks backend
servers:
  - url: http://localhost:7000
    description: Local development server
paths:
...
  /kid/filter:
    get:
      summary: Filter kids' products by category and/or color
      tags:
        - Kids
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
          example: "Clothing"
        - name: color
          in: query
          required: false
          schema:
            type: string
          example: "Red"
      responses:
        "200":
          description: Filtered list of kids' products
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                example:
                  - title: "Kids' T-Shirt"
                    price: 9.99
                    category: "Clothing"
                    color: "Red"
                    offer: true
...
  /kid/sort:
    get:
      summary: Sort kids' products by price or title
      tags:
        - Kids
      parameters:
        - name: sort
          in: query
          required: false
          schema:
            type: string
            enum:
              - lth
              - htl
              - asc
              - desc
          example: "lth"
      responses:
        "200":
          description: Sorted list of kids' products
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                example:
                  - title: "Kids' Toy"
                    price: 9.99
                    category: "Toys"
                    color: "Blue"
                    offer: false
  /kid/search:
    get:
      summary: Search for kids' products by title
      tags:
        - Kids
      parameters:
        - name: title
          in: query
          required: true
          schema:
            type: string
          example: "Toy"
      responses:
        "200":
          description: List of kids' products matching the search term
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                example:
                  - title: "Kids' Toy"
                    price: 19.99
                    category: "Toys"
                    color: "Blue"
                    offer: false
  /kid/create:
    post:
      summary: Create a new kids' product (admin only)
      tags:
        - Kids
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - price
                - category
                - color
                - offer
              properties:
                title:
                  type: string
                  example: "Kids' T-Shirt"
                price:
                  type: number
                  format: float
                  example: 15.99
                category:
                  type: string
                  example: "Clothing"
                color:
                  type: string
                  example: "Blue"
                offer:
                  type: boolean
                  example: false
      responses:
        "200":
          description: Product created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    example: "New product created"
  /kid/update/{id}:
    patch:
      summary: Update a kids' product (admin only)
      tags:
        - Kids
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: "60c72b2f9f1b2c3d8c1e1d1a"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: "Updated Kids' T-Shirt"
                price:
                  type: number
                  format: float
                  example: 12.99
                category:
                  type: string
                  example: "Clothing"
                color:
                  type: string
                  example: "Yellow"
                offer:
                  type: boolean
                  example: true
      responses:
        "200":
          description: Product updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    example: "Product updated"
  /kid/delete/{id}:
    delete:
      summary: Delete a kids' product (admin only)
      tags:
        - Kids
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: "60c72b2f9f1b2c3d8c1e1d1a"
      responses:
        "200":
          description: Product deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    example: "Product deleted"
  /cart:
    get:
      summary: Get all cart items for the authenticated user
      tags:
        - Cart
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of cart items for the authenticated user
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    productID:
                      type: string
                    quantity:
                      type: integer
                    price:
                      type: number
                      format: float
                    userID:
                      type: string
                example:
                  - productID: "60c72b2f9f1b2c3d8c1e1d1a"
                    quantity: 2
                    price: 19.99
                    userID: "60c72b2f9f1b2c3d8c1e1d1b"
                  - productID: "60c72b2f9f1b2c3d8c1e1d2a"
                    quantity: 1
                    price: 29.99
                    userID: "60c72b2f9f1b2c3d8c1e1d1b"
  /cart/create:
    post:
      summary: Add a product to the cart (authenticated user only)
      tags:
        - Cart
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - productID
                - quantity
                - userID
              properties:
                productID:
                  type: string
                  example: "60c72b2f9f1b2c3d8c1e1d1a"
                quantity:
                  type: integer
                  example: 1
                userID:
                  type: string
                  example: "60c72b2f9f1b2c3d8c1e1d1b"
      responses:
        "200":
          description: Product successfully added to cart
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    example: "Product added to Cart"
        "400":
          description: Product already in cart
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    example: "Product already available in Cart"
  /cart/update/{id}:
    patch:
      summary: Update a cart item (authenticated user only)
      tags:
        - Cart
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: "60c72b2f9f1b2c3d8c1e1d1a"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                quantity:
                  type: integer
                  example: 2
      responses:
        "200":
          description: Cart item updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    example: "Cart Updated"
        "403":
          description: Unauthorized access
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    example: "You are not authorized"
  /cart/delete/all:
    delete:
      summary: Remove all items from the cart (authenticated user only)
      tags:
        - Cart
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Cart emptied and items added to order history
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    example: "Cart Emptied"
        "403":
          description: Unauthorized access
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    example: "You are not authorized"
  /cart/delete/{id}:
    delete:
      summary: Delete a specific cart item (authenticated user only)
      tags:
        - Cart
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: "60c72b2f9f1b2c3d8c1e1d1a"
      responses:
        "200":
          description: Cart item deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    example: "Cart Item Deleted"
        "403":
          description: Unauthorized access
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    example: "You are not authorized"
...